fixed dunit failures

+++ geode-cq/src/distributedTest/java/org/apache/geode/cache/query/cq/dunit/CqQueryOptimizedExecuteDUnitTest.java
@@ -100,7 +100,7 @@ public class CqQueryOptimizedExecuteDUnitTest extends CqQueryDUnitTest {
       public void run2() throws CacheException {
         assertFalse("CqServiceImpl.EXECUTE_QUERY_DURING_INIT flag should be false ",
             CqServiceImpl.EXECUTE_QUERY_DURING_INIT);
-        int numOfQueryExecutions = (Integer) ((GemFireCacheImpl) getCache()).getCachePerfStats()
+        long numOfQueryExecutions = (Long) ((GemFireCacheImpl) getCache()).getCachePerfStats()
             .getStats().get("queryExecutions");
         assertEquals("Number of query executions for cq.execute should be 0 ", 0,
             numOfQueryExecutions);
@@ -190,7 +190,7 @@ public class CqQueryOptimizedExecuteDUnitTest extends CqQueryDUnitTest {
             CqServiceImpl.EXECUTE_QUERY_DURING_INIT);
         assertFalse(DistributionConfig.GEMFIRE_PREFIX + "cq.MAINTAIN_KEYS flag should be false ",
             CqServiceProvider.MAINTAIN_KEYS);
-        int numOfQueryExecutions = (Integer) ((GemFireCacheImpl) getCache()).getCachePerfStats()
+        long numOfQueryExecutions = (Long) ((GemFireCacheImpl) getCache()).getCachePerfStats()
             .getStats().get("queryExecutions");
         assertEquals("Number of query executions for cq.execute should be 0 ", 0,
             numOfQueryExecutions);
